Reorganize configs
