+++ configs/c2c_palette.py
