+++ configs/c2cbi_scz.py
